+++ fiboa_cli/datasets/commons/dictobject.py
@@ -1,0 +1,3 @@
+class DictObject(object):
+    def __init__(self, dict_):
+        self.__dict__.update(dict_)