+++ fiboa_cli/datasets/commons/ec.py
@@ -1,7 +1,7 @@
 import csv
 from io import StringIO
 from fiboa_cli.util import load_file
-
+from .dictobject import DictObject
 
 def add_eurocrops(base, year = None):
     if isinstance(base, dict):
@@ -54,10 +54,6 @@ In the data you'll find this as additional attributes:
 
     return ID, SHORT_NAME, TITLE, DESCRIPTION, PROVIDERS, EXTENSIONS, COLUMNS, LICENSE
 
-
-class DictObject(object):
-    def __init__(self, dict_):
-        self.__dict__.update(dict_)
 
 
 def load_ec_mapping(csv_file=None, url=None):