+++ fiboa_cli/datasets/nl_crop.py
@@ -1,4 +1,5 @@
 from ..convert_utils import convert as convert_
+from .commons.admin import add_admin
 import pandas as pd
 
 
@@ -77,6 +78,8 @@ MISSING_SCHEMAS = {
     }
 }
 
+COLUMNS, ADD_COLUMNS, EXTENSIONS = add_admin(vars(), "NL")
+
 
 def convert(output_file, cache = None, **kwargs):
     convert_(
@@ -88,7 +91,9 @@ def convert(output_file, cache = None, **kwargs):
         TITLE,
         DESCRIPTION,
         providers=PROVIDERS,
+        extensions=EXTENSIONS,
         missing_schemas=MISSING_SCHEMAS,
+        column_additions=ADD_COLUMNS,
         column_filters=COLUMN_FILTERS,
         column_migrations=COLUMN_MIGRATIONS,
         migration=MIGRATION,